+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Sections can be: Added Changed Deprecated Removed Fixed Security.
 
 ### Fixed
 
+- #474: PGN updates
 - #473: PGN 129546 field lengths and types
 - #472: PGN 127488 field Tilt/Trim is a percentage field